contract update

+++ assets/contracts/horcrux.sol
@@ -1,21 +1,35 @@
-pragma solidity ^ 0.5.0;
+pragma solidity ^0.5.0;
 
 interface IERC20 {
-    
-    function totalSupply() external view returns(uint256);
-    function balanceOf(address who) external view returns(uint256);
-    function allowance(address owner, address spender) external view returns(uint256);
-    function transfer(address to, uint256 value) external returns(bool);
-    function approve(address spender, uint256 value) external returns(bool);
-    function transferFrom(address from, address to, uint256 value) external returns(bool);
-    
+    function totalSupply() external view returns (uint256);
+
+    function balanceOf(address who) external view returns (uint256);
+
+    function allowance(address owner, address spender)
+        external
+        view
+        returns (uint256);
+
+    function transfer(address to, uint256 value) external returns (bool);
+
+    function approve(address spender, uint256 value) external returns (bool);
+
+    function transferFrom(
+        address from,
+        address to,
+        uint256 value
+    ) external returns (bool);
+
     event Transfer(address indexed from, address indexed to, uint256 value);
-    event Approval(address indexed owner, address indexed spender, uint256 value);
+    event Approval(
+        address indexed owner,
+        address indexed spender,
+        uint256 value
+    );
 }
 
 library SafeMath {
-
-    function mul(uint256 a, uint256 b) internal pure returns(uint256) {
+    function mul(uint256 a, uint256 b) internal pure returns (uint256) {
         if (a == 0) {
             return 0;
         }
@@ -24,23 +38,23 @@ library SafeMath {
         return c;
     }
 
-    function div(uint256 a, uint256 b) internal pure returns(uint256) {
+    function div(uint256 a, uint256 b) internal pure returns (uint256) {
         uint256 c = a / b;
         return c;
     }
 
-    function sub(uint256 a, uint256 b) internal pure returns(uint256) {
+    function sub(uint256 a, uint256 b) internal pure returns (uint256) {
         assert(b <= a);
         return a - b;
     }
 
-    function add(uint256 a, uint256 b) internal pure returns(uint256) {
+    function add(uint256 a, uint256 b) internal pure returns (uint256) {
         uint256 c = a + b;
         assert(c >= a);
         return c;
     }
 
-    function ceil(uint256 a, uint256 m) internal pure returns(uint256) {
+    function ceil(uint256 a, uint256 m) internal pure returns (uint256) {
         uint256 c = add(a, m);
         uint256 d = sub(c, 1);
         return mul(div(d, m), m);
@@ -48,34 +62,36 @@ library SafeMath {
 }
 
 contract ERC20Detailed is IERC20 {
-
     string private _name;
     string private _symbol;
     uint8 private _decimals;
 
-    constructor(string memory name, string memory symbol, uint8 decimals) public {
+    constructor(
+        string memory name,
+        string memory symbol,
+        uint8 decimals
+    ) public {
         _name = name;
         _symbol = symbol;
         _decimals = decimals;
     }
 
-    mapping(address => uint) balances;
+    mapping(address => uint256) balances;
 
-    function name() public view returns(string memory) {
+    function name() public view returns (string memory) {
         return _name;
     }
 
-    function symbol() public view returns(string memory) {
+    function symbol() public view returns (string memory) {
         return _symbol;
     }
 
-    function decimals() public view returns(uint8) {
+    function decimals() public view returns (uint8) {
         return _decimals;
     }
 }
 
 contract Owned {
-
     address public owner;
     address public newOwner;
 
@@ -103,7 +119,6 @@ contract Owned {
 }
 
 contract X_NSTests is ERC20Detailed, Owned {
-
     using SafeMath for uint256;
 
     mapping(address => uint256) private _balances;
@@ -111,36 +126,47 @@ contract X_NSTests is ERC20Detailed, Owned {
 
     string constant tokenName = "CRX5.0";
     string constant tokenSymbol = "CRx5";
-    uint8  constant tokenDecimals = 5;
+    uint8 constant tokenDecimals = 5;
     uint256 _totalSupply = 0;
-    
+
     // address of main contract
     address public authorizedContract;
-    
+
     // if true, owner cannot change authorized contract
     bool public contractIsUnchangable;
 
-    constructor() public payable ERC20Detailed(tokenName, tokenSymbol, tokenDecimals) {
+    constructor()
+        public
+        payable
+        ERC20Detailed(tokenName, tokenSymbol, tokenDecimals)
+    {
         _mint(msg.sender, _totalSupply);
     }
-    
+
     // return total supply of tokens
-    function totalSupply() public view returns(uint256) {
+    function totalSupply() public view returns (uint256) {
         return _totalSupply;
     }
 
     // query account balance for a address
-    function balanceOf(address owner) public view returns(uint256) {
+    function balanceOf(address owner) public view returns (uint256) {
         return _balances[owner];
     }
 
     // query allowed tokens for a address
-    function allowance(address owner, address spender) public view returns(uint256) {
+    function allowance(address owner, address spender)
+        public
+        view
+        returns (uint256)
+    {
         return _allowed[owner][spender];
     }
 
     // add main contract address, only callable once
-    function addAuthorizedContract(address main_contract_address) public returns(bool) {
+    function addAuthorizedContract(address main_contract_address)
+        public
+        returns (bool)
+    {
         require(msg.sender == owner);
         require(!contractIsUnchangable);
         authorizedContract = main_contract_address;
@@ -148,9 +174,9 @@ contract X_NSTests is ERC20Detailed, Owned {
         contractIsUnchangable = true;
         return true;
     }
-    
+
     // if conditions are met burn on transfer, else exicute a normal transfer
-    function transfer(address to, uint256 value) public returns(bool) {
+    function transfer(address to, uint256 value) public returns (bool) {
         require(value <= _balances[msg.sender]);
         require(to != address(0));
         _balances[msg.sender] = _balances[msg.sender].sub(value);
@@ -160,7 +186,7 @@ contract X_NSTests is ERC20Detailed, Owned {
     }
 
     // approve allowance
-    function approve(address spender, uint256 value) public returns(bool) {
+    function approve(address spender, uint256 value) public returns (bool) {
         require(spender != address(0));
         _allowed[msg.sender][spender] = value;
         emit Approval(msg.sender, spender, value);
@@ -168,7 +194,11 @@ contract X_NSTests is ERC20Detailed, Owned {
     }
 
     // transfer from approved tokens
-    function transferFrom(address from, address to, uint256 value) public returns(bool) {
+    function transferFrom(
+        address from,
+        address to,
+        uint256 value
+    ) public returns (bool) {
         require(value <= _balances[from]);
         require(value <= _allowed[from][msg.sender]);
         require(to != address(0));
@@ -181,17 +211,27 @@ contract X_NSTests is ERC20Detailed, Owned {
     }
 
     // increase msg.senders allowance
-    function increaseAllowance(address spender, uint256 addedValue) public returns(bool) {
+    function increaseAllowance(address spender, uint256 addedValue)
+        public
+        returns (bool)
+    {
         require(spender != address(0));
-        _allowed[msg.sender][spender] = (_allowed[msg.sender][spender].add(addedValue));
+        _allowed[msg.sender][spender] = (
+            _allowed[msg.sender][spender].add(addedValue)
+        );
         emit Approval(msg.sender, spender, _allowed[msg.sender][spender]);
         return true;
     }
 
     // decrease msg.senders allowance
-    function decreaseAllowance(address spender, uint256 subtractedValue) public returns(bool) {
+    function decreaseAllowance(address spender, uint256 subtractedValue)
+        public
+        returns (bool)
+    {
         require(spender != address(0));
-        _allowed[msg.sender][spender] = (_allowed[msg.sender][spender].sub(subtractedValue));
+        _allowed[msg.sender][spender] = (
+            _allowed[msg.sender][spender].sub(subtractedValue)
+        );
         emit Approval(msg.sender, spender, _allowed[msg.sender][spender]);
         return true;
     }
